requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
redis==5.0.1
cachetools==5.3.2
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.25.2
pydantic==2.4.2
python-multipart==0.0.6

# Supabase integration
supabase==2.3.0
postgrest-py==0.13.0

# Redis integration
upstash-redis==0.15.0
redis==5.0.1

# Async database support
asyncpg==0.29.0
databases[postgresql]==0.8.0

# Enhanced SQLAlchemy for Supabase
sqlalchemy[asyncio]==2.0.23
asyncio-mqtt==0.13.0

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Metrics
prometheus_client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
responses==0.24.1
